fix: move ChangePasswordForm to a component file in packages/components/src
